// web/app/components/workflow/types.ts
export type ValueSelector = string[] // [nodeId, key | obj key path]
// ...
export const VarType = {
  string: 'string',
  number: 'number',
  integer: 'integer',
  secret: 'secret',
  boolean: 'boolean',
  object: 'object',
  file: 'file',
  array: 'array',
  arrayString: 'array[string]',
  arrayNumber: 'array[number]',
  arrayObject: 'array[object]',
  arrayBoolean: 'array[boolean]',
  arrayFile: 'array[file]',
  any: 'any',
  arrayAny: 'array[any]',
} as const

export type VarType = (typeof VarType)[keyof typeof VarType]

export const ValueType = {
  variable: 'variable',
  constant: 'constant',
} as const

export type ValueType = (typeof ValueType)[keyof typeof ValueType]

export type Var = {
  variable: string
  type: VarType
  children?: Var[] | StructuredOutput // if type is obj, has the children struct
  isParagraph?: boolean
  isSelect?: boolean
  options?: string[]
  required?: boolean
  des?: string
  isException?: boolean
  isLoopVariable?: boolean
  nodeId?: string
  isRagVariable?: boolean
  schemaType?: string
}